spam_ham/src/requirements.txt
matplotlib==3.5.0
numpy==1.18.5
pandas==1.3.4
Pillow==9.0.1
scikit_learn==1.0.2
seaborn==0.11.2
tensorflow==2.7.0
visualkeras==0.0.2
